Refactor building the search query into a function
The intention of this change is first of all to remove the nested if
blocks, which I'm not a fan of. Secondly, it's to move the functionality
to a more reusable location, and not have it directly included with
other, not-quite-related code. Thirdly, it's to continue learning about
how Go works, which I'm comfortable in knowing that I have.